+++ s/maybe/index.barrel.ts
@@ -3,7 +3,6 @@ export * from "./fns/get.js"
 export * from "./fns/is-nay.js"
 export * from "./fns/is-yay.js"
 export * from "./fns/nay.js"
-export * from "./fns/problems.js"
 export * from "./fns/yay.js"
 
 export * from "./validators/validator.js"