+++ s/maybe/index.ts
@@ -1,5 +1,12 @@
 
 export * as maybe from "./index.barrel.js"
-export {yay, nay, isYay, isNay, gotYay, gotNay, validator, allow, deny} from "./index.barrel.js"
 export * from "./types.js"
+
+export {
+	yay, nay,
+	isYay, isNay,
+	getValue, gotValue,
+	getProblems, gotProblems,
+	validator, allow, deny,
+} from "./index.barrel.js"